main/requirements/base.txt
asgiref==3.5.1
backports.zoneinfo==0.2.1
Django==4.0.4
psycopg2==2.9.3
python-decouple==3.6
sqlparse==0.4.2
